+++ nzshm_model/psha_adapter/psha_adapter_interface.py
@@ -1,6 +1,6 @@
 import pathlib
 from abc import ABC, abstractmethod
-from typing import Union
+from typing import Dict, Union
 
 
 class PshaAdapterInterface(ABC):
@@ -12,9 +12,21 @@ class PshaAdapterInterface(ABC):
         self._source_logic_tree = source_logic_tree
 
     @abstractmethod
-    def fetch_resources(self, target_folder):
+    def fetch_resources(self, cache_folder):
+        """pull required data from api and store in target_folder"""
         pass
 
     @abstractmethod
-    def write_config(self, target_folder: Union[pathlib.Path, str]) -> pathlib.Path:
+    def unpack_resources(self, cache_folder: Union[pathlib.Path, str], target_folder: Union[pathlib.Path, str]):
+        """unpack resources from cache_folder to target folder returning mapping of unpacked filepaths"""
+        pass
+
+    @abstractmethod
+    def write_config(
+        self,
+        cache_folder: Union[pathlib.Path, str],
+        target_folder: Union[pathlib.Path, str],
+        resource_map: Dict[str, list[pathlib.Path]],
+    ) -> pathlib.Path:
+        """Build an openquake config from given input arguments"""
         pass